+++ src/forms/dynamic-forms/dynamic-form.component.ts
@@ -33,8 +33,6 @@ export class DynamicFormComponent implements OnInit {
   public ngOnInit() {
     this.form = this.ics.toFormGroup(this.inputs);
     this.form.valueChanges.subscribe((value) => {
-      // called when the value is updated
-      console.log(value);
     });
   }