+++ src/media-uploader/media-uploader.service.ts
@@ -6,26 +6,32 @@ import {
 } from 'rxjs/Observable';
 import {
     FileUploader,
-    FileItem
+    FileItem,
+    FileUploaderOptions
 } from 'ng2-file-upload';
 declare var API_URL: string;
 @Injectable()
 export class MediaUploaderService {
     public uploader: FileUploader;
+    public opts: FileUploaderOptions;
     public apiUrl: string;
     constructor() {
         //
     }
-    public config(config?: any) {
+    public config(config?: FileUploaderOptions) {
         if (config) {
-          //
+            this.uploader = new FileUploader(config);
         } else {
           const uploadApi = `${API_URL}/medias`;
           this.uploader = new FileUploader({
               url: uploadApi,
-              itemAlias: 'body'
+              itemAlias: 'body',
           });
         }
+        return this.uploader;
+    }
+    public getCurrentUploader() {
+        return this.uploader;
     }
     public uploadMedia(fileItem: FileItem) {
         const complete = new Observable((observer) => {
@@ -45,10 +51,10 @@ export class MediaUploaderService {
     }
     public uploadAll() {
         const complete = new Observable((observer) => {
-            observer.next();
-            this.uploader.onCompleteAll = (() => {
-                observer.next();
-            })
+            // observer.next();
+            this.uploader.onCompleteAll = () => {
+                observer.next({});
+            }
             /* uploader.onCompleteAll.onComplete = (
                 response: string,
                 status: number,